refactor(actions): Remove dependency cycle with actions/map.
