+++ lib/actions/form.js
@@ -25,6 +25,20 @@ export const setActiveSearch = createAction('SET_ACTIVE_SEARCH')
 export const clearDefaultSettings = createAction('CLEAR_DEFAULT_SETTINGS')
 export const storeDefaultSettings = createAction('STORE_DEFAULT_SETTINGS')
 
+/* SET_LOCATION action creator. Updates a from or to location in the store
+ *
+ * payload format: {
+ *   type: 'from' or 'to'
+ *   location: {
+ *     name: (string),
+ *     lat: (number)
+ *     lat: (number)
+ *   }
+ */
+
+// Private actions
+const settingLocation = createAction('SET_LOCATION')
+
 /**
  * Reset the trip form parameters to their default values.
  * @param {Boolean} [full=false] if set to true, the from/to locations and URL
@@ -215,5 +229,38 @@ export function formChanged(oldQuery, newQuery) {
       }
       debouncedPlanTrip()
     }
+  }
+}
+
+const clearingLocation = createAction('CLEAR_LOCATION')
+
+export function clearLocation(payload) {
+  return function (dispatch, getState) {
+    // Dispatch the clear location action and then clear the active search (so
+    // that the map and narrative are not showing a search when one or both
+    // locations are not defined).
+    dispatch(clearingLocation(payload))
+    dispatch(clearActiveSearch())
+  }
+}
+
+export function switchLocations() {
+  return async function (dispatch, getState) {
+    const { from, to } = getState().otp.currentQuery
+    // First, reverse the locations.
+    await dispatch(
+      settingLocation({
+        location: to,
+        locationType: 'from'
+      })
+    )
+    await dispatch(
+      settingLocation({
+        location: from,
+        locationType: 'to'
+      })
+    )
+    // Then kick off a routing query (if the query is invalid, search will abort).
+    return dispatch(routingQuery())
   }
 }
