+++ lib/actions/map.js
@@ -2,10 +2,6 @@ import { createAction } from 'redux-actions'
 import coreUtils from '@opentripplanner/core-utils'
 import getGeocoder from '@opentripplanner/geocoder'
 
-import { clearActiveSearch } from './form'
-import { deleteUserPlace } from './user'
-import { routingQuery } from './api'
-
 /* SET_LOCATION action creator. Updates a from or to location in the store
  *
  * payload format: {
@@ -18,33 +14,7 @@ import { routingQuery } from './api'
  */
 
 // Private actions
-const clearingLocation = createAction('CLEAR_LOCATION')
 const settingLocation = createAction('SET_LOCATION')
-const deleteRecentPlace = createAction('DELETE_LOCAL_USER_RECENT_PLACE')
-
-/**
- * Dispatches the action to delete a saved or recent place from localStorage.
- */
-export function forgetPlace(placeId, intl) {
-  return function (dispatch, getState) {
-    // localStorage only: Recent place IDs contain the string literal 'recent'.
-    if (placeId.indexOf('recent') !== -1) {
-      dispatch(deleteRecentPlace(placeId))
-    } else {
-      dispatch(deleteUserPlace(placeId, intl))
-    }
-  }
-}
-
-export function clearLocation(payload) {
-  return function (dispatch, getState) {
-    // Dispatch the clear location action and then clear the active search (so
-    // that the map and narrative are not showing a search when one or both
-    // locations are not defined).
-    dispatch(clearingLocation(payload))
-    dispatch(clearActiveSearch())
-  }
-}
 
 /**
  * Centers the given map to the coordinates of the specified place.
@@ -133,27 +103,6 @@ export function onLocationSelected(
     } else {
       dispatch(setLocation({ location, locationType }))
     }
-  }
-}
-
-export function switchLocations() {
-  return async function (dispatch, getState) {
-    const { from, to } = getState().otp.currentQuery
-    // First, reverse the locations.
-    await dispatch(
-      settingLocation({
-        location: to,
-        locationType: 'from'
-      })
-    )
-    await dispatch(
-      settingLocation({
-        location: from,
-        locationType: 'to'
-      })
-    )
-    // Then kick off a routing query (if the query is invalid, search will abort).
-    return dispatch(routingQuery())
   }
 }
