+++ lib/actions/ui.js
@@ -12,12 +12,17 @@ import {
 import { getModesForActiveAgencyFilter, getUiUrlParams } from '../util/state'
 import { getPathFromParts } from '../util/ui'
 
-import { clearActiveSearch, parseUrlQueryString, setActiveSearch } from './form'
-import { clearLocation, setMapCenter } from './map'
+import {
+  clearActiveSearch,
+  clearLocation,
+  parseUrlQueryString,
+  setActiveSearch
+} from './form'
 import { createOrUpdateUser } from './user'
 import { findRouteIfNeeded, findRoutesIfNeeded, setUrlSearch } from './api'
 import { MainPanelContent, MobileScreens } from './ui-constants'
 import { setActiveItinerary } from './narrative'
+import { setMapCenter } from './map'
 import { setRouterId } from './config'
 
 const updateLocale = createAction('UPDATE_LOCALE')
