+++ lib/actions/user.js
@@ -50,6 +50,7 @@ const deleteLocalUserSavedPlace = createAction('DELETE_LOCAL_USER_SAVED_PLACE')
 export const forgetStop = createAction('FORGET_STOP')
 export const rememberStop = createAction('REMEMBER_STOP')
 const rememberLocalUserPlace = createAction('REMEMBER_LOCAL_USER_PLACE')
+const deleteRecentPlace = createAction('DELETE_LOCAL_USER_RECENT_PLACE')
 
 function createNewUser(auth0User) {
   return {
@@ -772,6 +773,20 @@ export function rememberPlace(placeTypeLocation) {
       }
     } else if (persistenceMode.isLocalStorage) {
       dispatch(rememberLocalUserPlace(placeTypeLocation))
+    }
+  }
+}
+
+/**
+ * Dispatches the action to delete a saved or recent place from localStorage.
+ */
+export function forgetPlace(placeId, intl) {
+  return function (dispatch, getState) {
+    // localStorage only: Recent place IDs contain the string literal 'recent'.
+    if (placeId.indexOf('recent') !== -1) {
+      dispatch(deleteRecentPlace(placeId))
+    } else {
+      dispatch(deleteUserPlace(placeId, intl))
     }
   }
 }
