+++ lib/components/form/connected-location-field.js
@@ -9,6 +9,7 @@ import {
 import LocationField from '@opentripplanner/location-field'
 import styled from 'styled-components'
 
+import * as formActions from '../../actions/form'
 import * as mapActions from '../../actions/map'
 
 import connectLocationField from './connect-location-field'
@@ -51,7 +52,7 @@ const StyledLocationField = styled(LocationField)`
 
 export default connectLocationField(StyledLocationField, {
   actions: {
-    clearLocation: mapActions.clearLocation,
+    clearLocation: formActions.clearLocation,
     onLocationSelected: mapActions.onLocationSelected
   },
   includeLocation: true
