+++ lib/components/map/connected-endpoints-overlay.js
@@ -4,10 +4,11 @@ import { useIntl } from 'react-intl'
 import EndpointsOverlay from '@opentripplanner/endpoints-overlay'
 import React, { useCallback } from 'react'
 
-import { clearLocation, forgetPlace, setLocation } from '../../actions/map'
+import { clearLocation } from '../../actions/form'
+import { forgetPlace, rememberPlace } from '../../actions/user'
 import { getActiveSearch, getShowUserSettings } from '../../util/state'
 import { getUserLocations } from '../../util/user'
-import { rememberPlace } from '../../actions/user'
+import { setLocation } from '../../actions/map'
 
 const ConnectedEndpointsOverlay = (props) => {
   const intl = useIntl()
