+++ lib/index.js
@@ -39,7 +39,8 @@ import BatchSearchScreen from './components/mobile/batch-search-screen'
 import FormattedMode from './components/util/formatted-mode'
 import { setAutoPlan } from './actions/config'
 import { getCurrentPosition } from './actions/location'
-import { clearLocation, setLocationToCurrent, setMapCenter } from './actions/map'
+import { clearLocation } from './actions/form'
+import { setLocationToCurrent, setMapCenter } from './actions/map'
 import { findNearbyStops } from './actions/api'
 import createCallTakerReducer from './reducers/call-taker'
 import createOtpReducer from './reducers/create-otp-reducer'
